view for all events/actions the client sees
