+++ src/containers/GameClientEvents/index.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import View from './View';
 
 const mapStateToProps = (state) => ({
-  rawEvents: state.reducers.rawEvents,
+  gameClientEvents: state.reducers.gameClientEvents,
 });
 
 export default connect(mapStateToProps)(View);
